Move shared survey code to ui-core
