+++ ui-core/package.json
@@ -2,19 +2,33 @@
   "name": "@juniper/ui-core",
   "version": "0.0.1",
   "private": true,
+  "dependencies": {
+    "@fortawesome/free-solid-svg-icons": "6.2.x",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@tanstack/react-virtual": "^3.0.0-beta.54",
+    "classnames": "^2.3.2",
+    "downshift": "^7.6.0",
+    "inputmask": "3.3.10",
+    "lodash": "^4.17.21"
+  },
   "devDependencies": {
     "@babel/cli": "^7.21.5",
+    "@types/lodash": "^4.14.184",
     "@types/react": "^18.0.18"
   },
   "peerDependencies": {
-    "classnames": "^2.3.2",
     "react": "^18.2.0",
-    "react-router-dom": "^6.4.0"
+    "react-router-dom": "^6.4.0",
+    "survey-core": "1.9.75",
+    "survey-react-ui": "1.9.75",
+    "surveyjs-widgets": "1.9.75"
   },
   "main": "./build/index.js",
   "scripts": {
     "build": "NODE_ENV=production babel src --copy-files --extensions .js,.jsx,.ts,.tsx --out-dir build && tsc --declaration --emitDeclarationOnly",
-    "start": "chokidar 'src/**/*' -c 'npm run build'"
+    "start": "chokidar 'src/**/*' -c 'npm run build'",
+    "test": "react-scripts test",
+    "tsc": "tsc"
   },
   "eslintConfig": {
     "extends": "../.eslintrc.js"
