+++ ui-core/src/surveyjs/index.ts
@@ -1,0 +1,7 @@
+import './config'
+
+import './autosized-signaturepad'
+import './formatDate'
+import './inputmask'
+import './multiple-combobox-question'
+import './medications-question'
