+++ ui-core/src/surveyjs/inputmask.ts
@@ -1,0 +1,9 @@
+import 'inputmask/dist/inputmask/phone-codes/phone'
+
+import * as SurveyCore from 'survey-core'
+// eslint-disable-next-line
+// @ts-ignore
+import * as widgets from 'surveyjs-widgets'
+
+// See https://surveyjs.io/form-library/examples/control-data-entry-formats-with-input-masks/reactjs#content-code
+widgets.inputmask(SurveyCore)
