+++ ui-core/src/surveyjs/medications-question.ts
