+++ ui-core/src/surveyjs/medications.json
