+++ ui-core/src/types/forms.ts
@@ -1,3 +1,5 @@
+import { IPage, Question } from 'survey-core'
+
 export type VersionedForm = {
   id: string
   stableId: string
@@ -60,6 +62,36 @@ export type PreEnrollmentResponse = FormResponse & {
   answers: Answer[]
   fullData: string
   qualified: boolean
+}
+
+// Survey configuration
+
+/** these types are vague as we're still deciding how much custom stuff we need on top of SurveyJS */
+export type JuniperSurvey = {
+  pages: IPage[]
+  questionTemplates: JuniperQuestion[]
+}
+
+/** things that we need from SurveyJS elements to render the sheet view */
+export type ElementBase = {
+  name: string
+  type: string
+  title: string
+}
+
+/** Encompasses SurveyJS pages and panels. */
+export type ElementContainer = {
+  name: string
+  elements: ElementBase[]
+}
+
+/**
+ * We're extending SurveyJS to support templates -- the idea that a common question format may recur many times
+ * in a survey, and so should only be coded once
+ */
+export type JuniperQuestion = Question & {
+  questionTemplateName?: string
+  type: string
 }
 
 export {}
