+++ ui-participant/package.json
@@ -12,9 +12,7 @@
     "@tanstack/react-virtual": "^3.0.0-beta.54",
     "bootstrap": "^5.2.1",
     "classnames": "^2.3.2",
-    "downshift": "^7.6.0",
     "html-react-parser": "^3.0.4",
-    "inputmask": "3.3.10",
     "lodash": "^4.17.21",
     "micromark": "3.1.0",
     "oidc-client-ts": "^2.2.0",
@@ -43,7 +41,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "tsc": "tsc"
   },
   "browserslist": {
     "production": [
